#CE-393 add assignee to the entities
