--- Entity/Assignee.php
@@ -1,121 +1,0 @@
-<?php
-/*
- * This file is part of the CampaignChain package.
- *
- * (c) CampaignChain Inc. <info@campaignchain.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace CampaignChain\Hook\AssigneeBundle\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity
- * @ORM\Table(name="campaignchain_hook_assignee")
- */
-class Assignee
-{
-    /**
-     * @ORM\Column(type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    protected $entityId;
-
-    /**
-     * @ORM\Column(type="string")
-     */
-    protected $entityClass;
-
-    /**
-     * @ORM\ManyToOne(targetEntity="CampaignChain\CoreBundle\Entity\User")
-     */
-    protected $user;
-
-    /**
-     * Get id
-     *
-     * @return integer
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * Set entityId
-     *
-     * @param integer $entityId
-     * @return Assignee
-     */
-    public function setEntityId($entityId)
-    {
-        $this->entityId = $entityId;
-
-        return $this;
-    }
-
-    /**
-     * Get entityId
-     *
-     * @return integer
-     */
-    public function getEntityId()
-    {
-        return $this->entityId;
-    }
-
-    /**
-     * Set entityClass
-     *
-     * @param string $entityClass
-     * @return Assignee
-     */
-    public function setEntityClass($entityClass)
-    {
-        $this->entityClass = $entityClass;
-
-        return $this;
-    }
-
-    /**
-     * Get entityClass
-     *
-     * @return string
-     */
-    public function getEntityClass()
-    {
-        return $this->entityClass;
-    }
-
-    /**
-     * Set user
-     *
-     * @param \CampaignChain\CoreBundle\Entity\User $user
-     * @return Assignee
-     */
-    public function setUser(\CampaignChain\CoreBundle\Entity\User $user = null)
-    {
-        $this->user = $user;
-
-        return $this;
-    }
-
-    /**
-     * Get user
-     *
-     * @return \CampaignChain\CoreBundle\Entity\User
-     */
-    public function getUser()
-    {
-        return $this->user;
-    }
-}
