+++ Model/Assignee.php
@@ -1,0 +1,41 @@
+<?php
+/*
+ * This file is part of the CampaignChain package.
+ *
+ * (c) CampaignChain Inc. <info@campaignchain.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace CampaignChain\Hook\AssigneeBundle\Model;
+
+use CampaignChain\CoreBundle\Entity\User;
+
+class Assignee
+{
+    protected $user;
+
+    /**
+     * Set user
+     *
+     * @param User $user
+     * @return Assignee
+     */
+    public function setUser(User $user = null)
+    {
+        $this->user = $user;
+
+        return $this;
+    }
+
+    /**
+     * Get user
+     *
+     * @return User
+     */
+    public function getUser()
+    {
+        return $this->user;
+    }
+}
